Render text as svg, theme preview cleanup
