+++ frontend/components/ThemeGrid/ThemeGrid.styles.ts
@@ -4,7 +4,7 @@ import theme, { em } from '../../theme'
 export default {
   grid: css({
     display: 'flex',
-    justifyContent: 'space-between',
+    justifyContent: 'space-around',
     flexWrap: 'wrap',
     marginLeft: em(-theme.gutters.lg / 2),
     marginRight: em(-theme.gutters.lg / 2),
@@ -13,8 +13,9 @@ export default {
 
   item: css({
     width: '33.3333%',
+    // width: '50%',
     // width: '100%',
-    maxWidth: 440,
+    maxWidth: 460,
     paddingLeft: em(theme.gutters.lg / 2),
     paddingRight: em(theme.gutters.lg / 2),
     marginBottom: em(theme.gutters.lg),
