+++ frontend/components/ThemePreview/ActivityBar.tsx
@@ -2,6 +2,7 @@ import { Colors } from '@vscodethemes/types'
 import { css, cx } from 'emotion'
 import * as React from 'react'
 import Icon, { Icons } from '../Icon'
+import styles from './ActivityBar.styles'
 
 interface ActivityBarProps {
   colors: Colors
@@ -9,59 +10,35 @@ interface ActivityBarProps {
 
 const ActivityBar: React.SFC<ActivityBarProps> = ({ colors }) => (
   <div
-    className={classes.activityBar}
+    className={styles.activityBar}
     style={{ background: colors.activityBarBackground }}
   >
     <Icon
-      className={cx(classes.icon, classes.active)}
+      className={cx(styles.icon, styles.active)}
       icon={Icons.vscodeExplorer}
       fill={colors.activityBarForeground}
     />
     <Icon
-      className={classes.icon}
+      className={styles.icon}
       icon={Icons.vscodeSearch}
       fill={colors.activityBarForeground}
     />
     <Icon
-      className={classes.icon}
+      className={styles.icon}
       icon={Icons.vscodeGit}
       fill={colors.activityBarForeground}
     />
     <Icon
-      className={classes.icon}
+      className={styles.icon}
       icon={Icons.vscodeDebug}
       fill={colors.activityBarForeground}
     />
     <Icon
-      className={classes.icon}
+      className={styles.icon}
       icon={Icons.vscodeExtensions}
       fill={colors.activityBarForeground}
     />
   </div>
 )
-
-export const activityBarWidth = 8
-
-const classes = {
-  activityBar: css({
-    position: 'absolute',
-    top: 0,
-    left: 0,
-    width: `${activityBarWidth}%`,
-    height: '100%',
-    paddingTop: '1%',
-  }),
-
-  icon: css({
-    width: '100%',
-    height: '8%',
-    margin: '12% 0',
-    opacity: 0.5,
-  }),
-
-  active: css({
-    opacity: 1,
-  }),
-}
 
 export default ActivityBar
