+++ frontend/components/ThemePreview/Editor.tsx
@@ -1,22 +1,12 @@
 import { Colors } from '@vscodethemes/types'
 import { css } from 'emotion'
 import * as React from 'react'
-import { em } from '../../theme'
 
 interface EditorProps {
   colors: Colors
 }
 
-const Editor: React.SFC<EditorProps> = ({ colors, children }) => (
-  <div
-    className={classes.editor}
-    style={{ background: colors.editorBackground }}
-  >
-    {children}
-  </div>
-)
-
-const classes = {
+const styles = {
   editor: css({
     position: 'relative',
     top: 0,
@@ -25,11 +15,16 @@ const classes = {
     background: '#f0f0f0',
     // Fixed aspect ratio
     paddingTop: `${100 / (16 / 10)}%`,
-
-    [`@media (max-width: ${em(320)})`]: {
-      fontSize: '0.85rem',
-    },
   }),
 }
+
+const Editor: React.SFC<EditorProps> = ({ colors, children }) => (
+  <div
+    className={styles.editor}
+    style={{ background: colors.editorBackground }}
+  >
+    {children}
+  </div>
+)
 
 export default Editor
