+++ frontend/components/ThemePreview/StatusBar.tsx
@@ -1,7 +1,6 @@
 import { Colors } from '@vscodethemes/types'
-import { css, cx } from 'emotion'
 import * as React from 'react'
-import theme, { em } from '../../theme'
+import styles from './StatusBar.styles'
 
 interface StatusBarProps {
   colors: Colors
@@ -15,17 +14,17 @@ const StatusBar: React.SFC<StatusBarProps> = ({
   repositoryOwner,
 }) => (
   <div
-    className={classes.statusBar}
+    className={styles.statusBar}
     style={{ background: colors.statusBarBackground }}
   >
     {repositoryOwner && (
       <a
-        className={classes.link}
+        className={styles.link}
         href={`https://github.com/${repositoryOwner}/${repository}`}
         style={{ color: colors.statusBarForeground }}
       >
         <img
-          className={classes.pic}
+          className={styles.pic}
           src={`https://github.com/${repositoryOwner}.png?size=40`}
         />
         {repositoryOwner}
@@ -33,47 +32,5 @@ const StatusBar: React.SFC<StatusBarProps> = ({
     )}
   </div>
 )
-
-const statusBarGutter = theme.gutters.xs
-
-const classes = {
-  statusBar: css({
-    height: `${em(theme.fontSizes.xs + statusBarGutter * 2)}`,
-    borderBottomLeftRadius: em(theme.borderRadius.md),
-    borderBottomRightRadius: em(theme.borderRadius.md),
-    display: 'flex',
-    alignItems: 'center',
-  }),
-
-  link: css({
-    // flex: 1,
-    display: 'flex',
-    alignItems: 'center',
-    textDecoration: 'none',
-    fontSize: theme.fontSizes.xs,
-    height: '100%',
-  }),
-
-  pic: css({
-    width: em(theme.fontSizes.xs * 2),
-    height: em(theme.fontSizes.xs * 2),
-    marginLeft: em(theme.gutters.sm),
-    marginRight: em(theme.gutters.xs),
-    objectFit: 'cover',
-    borderRadius: '100%',
-  }),
-
-  secondary: css({
-    display: 'flex',
-    alignItems: 'center',
-    justifyContent: 'flex-end',
-  }),
-
-  icon: css({
-    height: em(theme.fontSizes.xs * 1.75),
-    marginLeft: em(theme.gutters.xs),
-    marginRight: `${statusBarGutter}%`,
-  }),
-}
 
 export default StatusBar
